+++ CHANGELOG.md
@@ -12,11 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Refactored Python and C++ unittests.
  - Added C# interface.
  - Moved documentation from readthedocs to github pages.
+ - Non-elementary checks for 2-cycles (`i->j->i` are not allowed)
 
 ### Added
  - [WIP] New functionality to select monotonic resources.
  - [WIP] Detection of monotonic resources.
- - Logger using [`sdplog`](https://github.com/gabime/spdlog)
+ - Logger using [`sdplog`](https://github.com/gabime/spdlog).
 
 ### Fixed
  - Issue: Bidirectional algorithm is not finding valid paths when using non-zero minimum resource values #89